move nucleus run trackers into one dir
