+++ src/main/kotlin/dev/nyon/skylper/skyblock/mining/hollows/tracker/nucleus/CrystalRunListener.kt
@@ -1,10 +1,11 @@
-package dev.nyon.skylper.skyblock.mining.hollows.locations
+package dev.nyon.skylper.skyblock.mining.hollows.tracker.nucleus
 
 import dev.nyon.skylper.extensions.*
 import dev.nyon.skylper.extensions.EventHandler.listenEvent
 import dev.nyon.skylper.minecraft
 import dev.nyon.skylper.skyblock.mining.hollows.Crystal
 import dev.nyon.skylper.skyblock.mining.hollows.HollowsModule
+import dev.nyon.skylper.skyblock.mining.hollows.locations.HollowsLocation
 
 object CrystalRunListener {
     private val crystalFoundRegex = regex("chat.hollows.run.crystalFound")
